Improve implementation details of getrf
